% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AncestRy_functions.R
\name{compute_introgression_lengths}
\alias{compute_introgression_lengths}
\title{Length of Introgressed Sequences}
\usage{
compute_introgression_lengths(
  x,
  pop = 1,
  allele = c(native = 1, introgressed = 2),
  loc.pos,
  stats = NULL,
  exclude = T,
  pool = F
)
}
\arguments{
\item{x}{file output from \code{evolve2.0} or \code{evolve2.1} simulations in (\code{struct} format).}

\item{pop}{[numeric] Which simulated population should the length of introgressed sequences be estimated on.}

\item{allele}{[numeric] A vector describing (in order) the integer for the native and introgressed allele.}

\item{loc.pos}{[numeric] A vector describing the position of each marker.}

\item{stats}{[character] One of "length" or "nb.loci". Specifying this argument triggers the estimation of pooled individuals mean, median, mode(s), and frequencies of introgressed sequences length (in bp --> "length" or number of loci --> "nb.loci").}

\item{exclude}{[logical] Whether or not introgressed sequences of length = 1bp or 1 locus should be recorded (TURE = no, FALSE = yes).}

\item{pool}{[logical] Whether results should be pooled together (one vector with all length across all homologs and individuals).}
}
\description{
This function computes the length of introgressed sequences within a select population. Length is calculated for each individual and chromosome within individuals separately (but see arguments \code{pool} and \code{stats} for other outputs).
}
