% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AncestRy_functions.R
\name{compute_introgression_peaks}
\alias{compute_introgression_peaks}
\title{Introgression Peaks (Islands or Valleys of Introgression)}
\usage{
compute_introgression_peaks(
  x,
  pop = 1,
  introgressed.allele = 2,
  loc.pos,
  cores = 1,
  n.perm = 999,
  exclude = T
)
}
\arguments{
\item{x}{[list] file output from \code{evolve2.0} or \code{evolve2.1} simulations in (\code{struct} format).}

\item{pop}{[numeric] Which simulated population should the peaks be calculated on.}

\item{introgressed.allele}{[numeric] The number representing the introgressed allele.}

\item{loc.pos}{[numeric] A vector describing the position of each marker.}

\item{cores}{[numeric] Number of cores to use for HMM analysis.}

\item{n.perm}{[numeric] N number of permutations to assess significance of island and valleys of introgression inferred with HMM.}

\item{exclude}{[logical] Whether island and valleys of length 1 should be kept (FALSE) or not (TRUE).}
}
\description{
This function takes the output of \code{\link{evolve2.0}} or \code{\link{evolve2.1}} and assess peaks of high (island) and low (valley) introgression. Peaks are assessed based on posterior probabilities of states inferred using a hidden markov model and permutation test (3 states: Background, high, low). This function also computes the length of islands and valleys of introgression as well as the distance between them.
}
\note{
The function \code{HMM} and \code{perHMM} are for internal use only.
}
