% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AncestRy_functions.R
\name{create.events}
\alias{create.events}
\title{Create Historical Events}
\usage{
create.events(nb.events, generation, migration.rate.switch, write = F)
}
\arguments{
\item{nb.events}{[numeric] The number of historical events.}

\item{generation}{[numeric] Generations you wish to change the admixture rate between populations.}

\item{migration.rate.switch}{[numeric] The index (number) of the matrix you wish to use at a select generation (see \code{\link{evolve2.0}} or \code{\link{evolve2.1}} for details).}

\item{write}{[logical] If true, the events file will be saved to the working directory.}
}
\value{
This function returns a data frame with two columns (Generation and migration.rate) and a number of rows equal to the number of historical events. The column Generation record the generation at which one want to modify the migration between populations. The column migration.rate record which of the migration matrices would like to used from that generation forward. See \code{\link{evolve2.0}} or \code{\link{evolve2.1}}.
}
\description{
This function generates the \code{events} data frame needed to run \code{\link{evolve2.0}} and \code{\link{evolve2.1}}.
}
