% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AncestRy_functions.R
\name{create.source}
\alias{create.source}
\title{Create a Source Population}
\usage{
create.source(N, nl, na, pop.index)
}
\arguments{
\item{N}{[numeric] Number of individuals in the initial population.}

\item{nl}{[numeric] Number of simulated loci.}

\item{na}{[numeric] Number of alleles at each locus.}
}
\description{
This function allows you to create a \code{source} population for genetic simulation using \code{\link{evolve2.0}} or \code{\link{evolve2.1}}. Specifically, using the argument \code{source}, one can define a population serving indefinitely as a source of individuals. This function generates the value that needs to be fed to the \code{source} argument in \code{evolve2.0} or \code{evolve2.1} if such a demographic scenario wishes to be simulated.
}
