% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AncestRy_functions.R
\name{evolve2.0}
\alias{evolve2.0}
\alias{evolve2.1}
\title{Evolution of Genetic and Genomic Landscape: \code{\link{evolve2.0}} and \code{\link{evolve2.1}}}
\usage{
evolve2.0(
  x,
  time,
  type = c("constant", "dynamic", "additive", "custom"),
  recombination = c("map", "average"),
  recom.rate,
  loci.pos = NULL,
  chromo_mb = NULL,
  init.sex = NULL,
  migration.rate.list = NULL,
  migration.rate.initial = NULL,
  mutation.rate = NULL,
  param.z = NULL,
  param.w = NULL,
  fun = c(phenotype = NULL, fitness = NULL),
  events = NULL,
  gen.snapshot = NULL,
  source = NULL,
  logfile = FALSE
)

evolve2.1(
  x,
  time,
  type = c("constant", "dynamic", "additive", "custom"),
  recombination = c("map", "average"),
  recom.rate,
  loci.pos = NULL,
  chromo_mb = NULL,
  init.sex = NULL,
  migration.rate.list = NULL,
  migration.rate.initial = NULL,
  mutation.rate = NULL,
  param.z = NULL,
  param.w = NULL,
  fun = c(phenotype = NULL, fitness = NULL),
  events = NULL,
  gen.snapshot = NULL,
  source = NULL,
  logfile = FALSE
)
}
\arguments{
\item{migration.rate.list}{[list] A list of square matrices defining the migration rate between all populations simulated. The index of these matrices (1,2,3,...) is used in the events file (see \code{\link{create.events}}) will you these indices to change the migration rate for desired generations.}

\item{migration.rate.initial}{[numeric] The index of the matrix that wish to be used before the occurrence of the first historical event.}

\item{events}{The \code{events} data frame generated using \code{create.events function}.}

\item{gen.snapshot}{[numeric] A vector of generations at which you want data to be output.}

\item{source}{[list]  A list describing the structure of the source population (see \code{\link{create.source}}). As currently implemented, only the first population of a larger metapopulation can serve as an infinite source of individuals.}

\item{logfile}{[logical] Whether or not a log file recording historical events should be generated.}
}
\description{
These functions are identical to the \code{evolve} in \code{glads}, except that it allows to vary (change) the migration rates over time by defining a data frame describing the desired migration rate at a select generation (equivalent to historical events in fastsimcoal). It also allows select demes to start empty as well as to recurrently use a source population when simulating under the "constant" model for many generations. Finally, these functions allow to retrieve simulated data at user-defined generations to evaluate temporal changes in populations' genetic structure.
}
\note{
\code{evolve2.0} works with \code{glads 0.1.1}.
\code{evolve2.1} works with \code{glads 0.1.2}.
}
\seealso{
\code{help(evolve)} for additional details on the function arguments.
}
