% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AncestRy_functions.R
\name{homogeneous.struct}
\alias{homogeneous.struct}
\title{Homogeneous Initial Structure}
\usage{
homogeneous.struct(N, nl, n_pop = 1, allele.code = 2)
}
\arguments{
\item{N}{[numeric] Number of individuals within initial populations.}

\item{nl}{[numeric] Number of simulated loci.}

\item{n_pop}{[numeric] Number of populations with the same initial structure to create.}

\item{allele.code}{[numeric] The code to give to the fixed allele within populations.}
}
\description{
This function allows to set an initial population structure for which all individuals within populations are fixed for a given allele.
}
