% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AncestRy_functions.R
\name{iterate_evolve2.0}
\alias{iterate_evolve2.0}
\title{Wrap up function: Perform Multiple Simulations}
\usage{
iterate_evolve2.0(
  x,
  time,
  type = c("constant", "dynamic", "additive", "custom"),
  recombination = c("map", "average"),
  recom.rate,
  loci.pos = NULL,
  chromo_mb = NULL,
  init.sex = NULL,
  migration.rate.list = NULL,
  migration.rate.initial = NULL,
  mutation.rate = NULL,
  param.z = NULL,
  param.w = NULL,
  fun = c(phenotype = NULL, fitness = NULL),
  events = NULL,
  gen.snapshot = NULL,
  source = NULL,
  logfile = FALSE,
  n_rep = 100,
  nb.cores = 1
)
}
\arguments{
\item{n_rep}{[numeric]  Number of simulations to be performed using \code{evolve2.0} or \code{evolve2.1}.}

\item{nb.cores}{[numeric] Number of cores to use for simulation.}
}
\description{
These functions simply allow the user to run a user-defined number of simulations using \code{\link{evolve2.0}} or \code{\link{evolve2.1}}.
}
\seealso{
For details on other arguments see \code{\link{evolve2.0}} or \code{\link{evolve2.1}}.
}
