% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AncestRy_functions.R
\name{iterate_evolve2.1}
\alias{iterate_evolve2.1}
\title{#' @rdname iterate_evolve2.0}
\usage{
iterate_evolve2.1(
  x,
  time,
  type = c("constant", "dynamic", "additive", "custom"),
  recombination = c("map", "average"),
  recom.rate,
  loci.pos = NULL,
  chromo_mb = NULL,
  init.sex = NULL,
  migration.rate.list = NULL,
  migration.rate.initial = NULL,
  mutation.rate = NULL,
  param.z = NULL,
  param.w = NULL,
  fun = c(phenotype = NULL, fitness = NULL),
  events = NULL,
  gen.snapshot = NULL,
  source = NULL,
  logfile = FALSE,
  n_rep = 100,
  nb.cores = 1
)
}
\description{
#' @rdname iterate_evolve2.0
}
