% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AncestRy_functions.R
\name{perHMM}
\alias{perHMM}
\title{Randomization Test Following Hidden Markov Assessment}
\usage{
perHMM(x, iter = 9999, island, valley, exclude = T)
}
\description{
Randomization Test Following Hidden Markov Assessment
}
\keyword{internal}
